console: Refactor application collaborator add view
Move redux to 'connect.js' file
